Paper, Rock, Scissors Game in Java
In the kids’ game “Paper, Rock, Scissor” each player chooses one of the three and the winner is determined by the relationship
between the two choices. “Paper covers rock” so paper wins; “rock breaks scissor” so rock wins; and “scissor cuts paper” so
scissor wins. If the both choose the same, it is a tie and no one wins.
A Java program to play the game against the computer until the player enters "q" instead of a choice.
A static method machineChoice() that randomly generates a number (1, 2 or 3) and returns a character (p, r, s)
according to the number. 1 is for p, 2 is for r, 3 is for s.