Array Usage in Java (For Beginners)
menu(): display the menu on the screen, then read, validate and return the user’s choice.
initialize(): sets all integer elements of the array to zero.
fillArray(): reads the integer values from the user until zero is entered or the array size is reached. The method must
return the number of actual elements inputted as well as the array itself.
displayArray(): receives an integer array and the number of actual elements as parameters. Display all array elements
on the screen.
findPos(): searches for an integer value in the array and returns that position (index) of the first occurrence of the value
and returns -1 if the value is not present.